Allow processing service protocol callbacks on the thread of a root isolate.
The primary change is not holding the handlers lock while iterating over the views. This includes making a copy of the handlers before using the same. The rest of the change is dealing with the edge case of the shell being collected between the handler copy and use.
